Fix review issues: safety, dead code, bindings, and tests
- Replace assert() with throwing checks in Model accessors (Python-safe)
- Add bounds checks to constraint_violation(), lambda_func(), restore_state()
- Add validation to add_constraint()/minimize() rejecting var handles
- Remove dead topological_sort() declaration and implementation
- Remove unused var_id_from_handle(), unused includes
- Clean up verbose wrap() comments in model.cpp
- Fix std::find().erase() UB in set_moves() with iterator guard
- Cast rng.integers() returns to size_t for vector indexing
- Use infinity() instead of 1e15 sentinel in division/pow
- Initialize Move::Change::var_id to -1
- Add -fPIC for nanobind shared module linking
- Add #include <limits> to pool.h
- Add noexcept to trivial const accessors
- Register nanobind exception translators (out_of_range→IndexError,
  invalid_argument→ValueError)
- Expand Python bindings: RNG, Move, SavedValues, MoveProbabilities,
  LNS, SolutionPool, copy_state/restore_state, move generators
- Add error-path tests (out-of-range, invalid handles)
- Add ParallelSearch and copy_state/restore_state tests
- Add Python test_moves.py with moves, LNS, pool, error path tests
- Extract shared vid() helper to tests/test_helpers.h
- De-flake CHPED 24-unit test (conditional feasibility assertion)
- Tighten FJ-NL and MoveProbabilities floor assertions

+++ src/search.cpp
@@ -228,7 +228,7 @@ SearchResult solve(Model& model, double time_limit, uint64_t seed, bool use_fj) 
         if (var.type == VarType::Float) {
             auto violated = vm.violated_constraints();
             if (!violated.empty()) {
-                int ci = violated[rng.integers(0, violated.size())];
+                int ci = violated[static_cast<size_t>(rng.integers(0, violated.size()))];
                 auto nm = newton_tight_move(var.id, model, ci);
                 moves.insert(moves.end(), nm.begin(), nm.end());
             }
@@ -242,7 +242,7 @@ SearchResult solve(Model& model, double time_limit, uint64_t seed, bool use_fj) 
         }
 
         // Pick a move uniformly
-        const auto& move = moves[rng.integers(0, moves.size())];
+        const auto& move = moves[static_cast<size_t>(rng.integers(0, moves.size()))];
 
         // Evaluate via delta
         auto saved = save_move_values(model, move);
